// constants/api.d.ts
export interface components {
    schemas: {
// ...
        Resturant: {
            readonly id: number;
            /** Format: uri */
            name: string;
            logo: string;
            poster: string;
            about?: string;
            /** Format: email */
            email: string | null;
            phone_number?: string | null;
            rating?: number;
            like_count?: number;
            review_count?: number;
            estimated_time?: number;
            delivery_fee?: number;
        };
// ...
        Dish: {
            readonly id: number;
            /** Format: uri */
            image: string;
            name: string;
            price: number;
            discount?: number;
            contents?: string[];
            description: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}